requirements.txt
Django
DjangoRestFramework
django-cors-headers